# Conflicts:
#	mobile/apps/photos/lib/ui/rituals/start_new_ritual_card.dart
#	mobile/apps/photos/lib/ui/viewer/search_tab/contacts_section.dart
#	mobile/apps/photos/lib/ui/viewer/search_tab/locations_section.dart

+++ mobile/apps/locker/lib/ui/pages/onboarding_page.dart
@@ -172,7 +172,7 @@ class _OnboardingPageState extends State<OnboardingPage> {
                           const SizedBox(height: 12),
                           DotsIndicator(
                             dotsCount: _featureCount,
-                            position: _activeDotIndex,
+                            position: _activeDotIndex.toDouble(),
                             animate: true,
                             animationDuration: const Duration(
                               milliseconds: 300,
